Change dependency gathering to ResolvedArtifict
The use of ModuleVersionIdentifier loses information from the original ResolvedArtifact, particularly, the file associated with the artifact. This file could be useful in resolving library name, version, and license information, and would thus need to be required.

Modified dependency gathering to return the set of ResolvedArtifacts (which probably doesn't even require a separate class).

Also:
* DependencyCollector.kt
  * Combine the two functions into a single, since no longer much logic in either.
  * Fix bug where all resolved artifacts are re-processed for every dependency in the configuration