Fix strings before writing XML
The `fixString()` method is called on items as soon as they are read, then characters that require special handling in XML and Android resource files are escaped immediately.  This makes them more difficult to process as their original values, and causes problems for other output formats.

Moved `fixString()` to the `AboutLibrariesTask` and renamed as `encodeForAndroidResourceValue()`, and call it before writing Android resource file.

Replaced processor class use of `fixString()` with `nullAsEmpty()` which converts `null` values to an `""` empty string.